requirements.txt
pyrogram
tgcrypto
groq
pymongo
python-dotenv
flask